replace siteMetadata
